CRYP-4449: Move onAssetsConversion from index.ts to onAssetsConversion.ts
